Split repository.go and repository_test.go by repository format
