+++ repository_apt_test.go
@@ -1,0 +1,49 @@
+package client
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func getTestRepositoryAptHosted(name string) Repository {
+	return Repository{
+		Name:   name,
+		Online: true,
+		Type:   RepositoryTypeHosted,
+		Format: RepositoryFormatApt,
+
+		RepositoryApt: &RepositoryApt{
+			Distribution: "bionic",
+		},
+		RepositoryAptSigning: &RepositoryAptSigning{
+			Keypair:    "string",
+			Passphrase: "string",
+		},
+		RepositoryCleanup: &RepositoryCleanup{
+			PolicyNames: []string{"weekly-cleanup"},
+		},
+		RepositoryStorage: &RepositoryStorage{
+			BlobStoreName:               "default",
+			StrictContentTypeValidation: true,
+			WritePolicy:                 "ALLOW",
+		},
+	}
+}
+
+func TestRepositoryAptHosted(t *testing.T) {
+	client := NewClient(getDefaultConfig())
+	repo := getTestRepositoryAptHosted("test-apt-repo-hosted")
+
+	err := client.RepositoryCreate(repo)
+	assert.Nil(t, err)
+
+	updatedRepo := repo
+	updatedRepo.Online = false
+
+	err = client.RepositoryUpdate(repo.Name, updatedRepo)
+	assert.Nil(t, err)
+
+	err = client.RepositoryDelete(repo.Name)
+	assert.Nil(t, err)
+}
