+++ repository_bower.go
@@ -1,0 +1,6 @@
+package client
+
+// RepositoryBower contains data of bower repositories
+type RepositoryBower struct {
+	RewritePackageUrls bool `json:"rewritePackageUrls"`
+}
