+++ repository_pypi.go
@@ -1,0 +1,3 @@
+package client
+
+// So far there is nothing special for PyPi repositories
